move tests, add test_identity
